Guard removal ov xz to being installed for boostrap rerun

+++ bootstrap.sh
@@ -110,7 +110,9 @@ if [ "$os" = "darwin" ]; then
 else
   # Remove liblzma installed from the_silver_searcher to avoid conflict with
   # system
-  brew remove xz
+  if brew list | grep -iq xz; then
+    brew remove xz
+  fi
 fi
 
 # Make use of newly installed ctags to index ruby
